+++ Basics - Core concept/07-cmp-communication-assignment-test/src/App.vue
@@ -1,0 +1,23 @@
+<template>
+  <user-data @add-contact="addContact" />
+  <active-user v-bind="person" />
+</template>
+
+<script>
+export default {
+  data() {
+    return {
+      person: {
+        personName: '',
+        personAge: '',
+      },
+    };
+  },
+  methods: {
+    addContact(name, age) {
+      this.person.personName = name;
+      this.person.personAge = age;
+    },
+  },
+};
+</script>