+++ Basics - Core concept/08-cmp-advanced/src/components/UI/BaseBadge.vue
@@ -1,0 +1,44 @@
+<template>
+  <!-- <span @click="selectBadge(type)" class="badge" :class="classes">{{
+    caption
+  }}</span> -->
+  <button @click="selectBadge(type)" class="badge" :class="classes">
+    {{ caption }}
+  </button>
+</template>
+
+<script>
+export default {
+  props: ['type', 'caption'],
+  emits: ['select-badge'],
+  inject: ['selectBadge'],
+  computed: {
+    classes() {
+      return {
+        'badge--admin': this.type === 'admin',
+        'badge--author': this.type === 'author',
+      };
+    },
+  },
+};
+</script>
+
+<style>
+.badge {
+  display: inline-block;
+  padding: 0.5rem 1rem;
+  border-radius: 30px;
+  background-color: #ccc;
+  color: #2e2e2e;
+}
+
+.badge--admin {
+  background-color: #810036;
+  color: white;
+}
+
+.badge--author {
+  background-color: #002c8a;
+  color: white;
+}
+</style>